Analytics/requirements.txt
# MMM Bayesian Analytics — Dependencies
# Install: pip install -r requirements.txt

# Core scientific stack
numpy>=1.24
pandas>=2.0
scipy>=1.11

# Bayesian inference
pymc>=5.10
arviz>=0.17
pytensor>=2.18

# Machine learning utilities
scikit-learn>=1.3

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7
seaborn>=0.13

# CLI
argparse
